MarketDataSimulator/requirements.txt
flask==2.3.2
kafka-python==2.0.2
python-dotenv==1.0.0  # Optional, for managing environment variables
pytest==7.4.0         # Optional, for testing